requirements.txt
tkinter
cryptography
Pillow